Split ssd.py into different python files in source folder
